requirements.txt
numpy==1.21.2
albumentations==1.0.3
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
Cython==0.29.24
decorator==4.4.2
dlib==19.21.0
easydict==1.9
face-recognition==1.3.0
face-recognition-models==0.3.0
faiss-cpu==1.6.3
Flask==1.1.2
flatbuffers==2.0
gevent==20.6.2
graphviz==0.8.4
greenlet==0.4.16
grequests==0.6.0
idna==2.10
imageio==2.9.0
importlib-metadata==4.8.1
insightface==0.4.2
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.2
networkx==2.5
onnx==1.10.1
onnxruntime==1.8.1
opencv-python==4.4.0.42
opencv-python-headless==4.5.3.56
pandas==1.1.2
Pillow==7.2.0
prettytable==2.2.0
protobuf==3.17.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.24.0
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
threadpoolctl==2.1.0
tifffile==2020.9.3
tqdm==4.49.0
typing-extensions==3.10.0.2
urllib3==1.25.10
wcwidth==0.2.5
Werkzeug==1.0.1
zipp==3.5.0
zope.event==4.5.0
zope.interface==5.1.0
